% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_sampling_load.R
\name{cs_sampling_load}
\alias{cs_sampling_load}
\title{Adjust draws from an already fitted Stan model}
\usage{
cs_sampling_load(
  svydes,
  fit,
  par_stan = NA,
  data_stan = NULL,
  rep_design = FALSE,
  ctrl_rep = list(replicates = 100, type = "mrbbootstrap"),
  H_estimate = "MCMC",
  matrix_sqrt = "eigen",
  prior_only = FALSE,
  yj_transform = TRUE,
  yj_args = list()
)
}
\arguments{
\item{svydes}{\itemize{
\item a \code{\link[survey]{svydesign}} object or a \code{\link[survey]{svrepdesign}} object. This contains cluster ID, strata, and weight information (\code{\link[survey]{svydesign}}) or replicate weight information (\code{\link[survey]{svrepdesign}})
}}

\item{fit}{\itemize{
\item a \code{\link[brms]{brmsfit}} or \code{\link[rstan]{stanfit}} object
}}

\item{par_stan}{\itemize{
\item a list of a subset of parameters to output after adjustment. All parameters are adjusted including the derived parameters, so users may want to only compare subsets. The default, NA, will return all parameters.
}}

\item{data_stan}{\itemize{
\item a list of data inputs used to construct \code{fit}. This is extracted automatically for a \code{\link[brms]{brmsfit}} object, but is required for a \code{\link[rstan]{stanfit}}.
}}

\item{rep_design}{\itemize{
\item logical indicating if the svydes object is a \code{\link[survey]{svrepdesign}}. If FALSE, the design will be converted to a \code{\link[survey]{svrepdesign}} using ctrl_rep settings
}}

\item{ctrl_rep}{\itemize{
\item a list of settings when converting svydes from a \code{\link[survey]{svydesign}} object to a \code{\link[survey]{svrepdesign}} object. replicates - number of replicate weights. type - the type of replicate method to use, the default is mrbbootstrap which sample half of the clusters in each strata to make each replicate (see \code{\link[survey]{as.svrepdesign}}).
}}

\item{H_estimate}{\itemize{
\item a string indicating the method to use to estimate H. The default "MCMC" is Monte Carlo averaging over posterior draws. Otherwise, a plug-in using the posterior mean.
}}

\item{matrix_sqrt}{\itemize{
\item a string indicating the method to use to take the "square root" of the R1 and R2 matrices. The default "eigen" uses the eigenvalue decomposition. Otherwise, the Cholesky decomposition is used.
}}

\item{prior_only}{\itemize{
\item a logical indicating if the stan model has an option for sampling just from the prior distribution. This can be used to further refine the estimates for covariances H and J.
}}

\item{yj_transform}{\itemize{
\item a logical indicating whether to use the Yeo-Johnson transformation. If \code{TRUE}, H is estimated from the transformed posterior covariance.
}}

\item{yj_args}{\itemize{
\item a named list of arguments pass to \code{\link[csSampling]{cs_sampling_yj}}. Including \code{diag_only}, \code{subset_matrix}, \code{export_unconst_pars}, and \code{yj_range}.
}}
}
\value{
The output of cs_sampling or cs_sampling_yj.
}
\description{
\code{cs_sampling_load} is a wrapper function that passes an existing \code{\link[brms]{brmsfit}} or \code{\link[rstan]{stanfit}} object to \code{\link[csSampling]{cs_sampling}} or, if \code{yj_transform} is \code{TRUE}, to \code{\link[csSampling]{cs_sampling_yj}}.
}
\examples{

# Survey design information
library(survey)
data(api)
apistrat$wt <- apistrat$pw / mean(apistrat$pw)

dstrat <- svydesign(ids = ~1, strata = ~stype, weights = ~wt, data = apistrat, fpc=~fpc)

# Fit a weighted brms model
library(brms)

fit <- brm(
    bf(api00|weights(wt) ~ ell + meals + mobility, center = FALSE),
    data = apistrat,
    family = gaussian(),
    save_pars = save_pars(all = TRUE)
)

# Apply the adjustment with Yeo-Johnson transformation 
adjusted_yj <- cs_sampling_load(svydes = dstrat, fit = fit)

# Apply the adjustment without Yeo-Johnson transformation 
adjusted <- cs_sampling_load(
    svydes = dstrat, fit = fit, yj_transform = FALSE
)

}
