% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat14}
\alias{dat14}
\title{2014 National Survey on Drug Use and Health Public Use File}
\format{
\subsection{\code{dat14}}{

A data frame with 55271 rows and 12 columns:
\describe{
\item{QUESTID2}{Respondent Identification}
\item{CIGMON}{Cigarettes - Past Month Use}
\item{AMDEY2_U}{Adult Past Year Major Depressive Episode}
\item{IRCIGYFU, CIGYFU}{Imputation-Revised and Recoded Year Of First Cigarette Use}
\item{IRMJYFU, MJYFU}{Imputation-Revised and Recoded Year Of First Marijuana/hashish Use}
\item{CATAG3, CATAG6}{Age Category Recode (5 Levels and 6 Levels)}
\item{ANALWT_C}{Final Personal-Level Sample Weight}
\item{VESTR}{Analysis Stratum}
\item{VEREP}{Analysis Replicate}
}
}
}
\source{
\url{https://www.datafiles.samhsa.gov/dataset/national-survey-drug-use-and-health-2014-nsduh-2014-ds0001}
}
\usage{
dat14
}
\description{
A subset of variables from the 2014 National Survey on Drug Use and Health Public Use File
}
\keyword{datasets}
