% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_sampling.r
\name{DEadj}
\alias{DEadj}
\title{DEadj}
\usage{
DEadj(par, par_hat, R2R1)
}
\arguments{
\item{par}{\itemize{
\item a 1D array/matrix of parameter values to be rescaled/rotated
}}

\item{par_hat}{\itemize{
\item a 1D array/matrix of center values, typically the mean
}}

\item{R2R1}{\itemize{
\item the rotation matrix to rescale/rotate par
}}
}
\value{
A 1D matrix corresponding to the input par that has been rescaled/rotated
}
\description{
DEadj
}
