% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_sampling.r
\name{grad_par}
\alias{grad_par}
\title{grad_par}
\usage{
grad_par(pwts, svydata, stanmod, standata, par_hat)
}
\arguments{
\item{pwts}{\itemize{
\item is the weights argument. withReplicates updates this for each set of replicate weights
}}

\item{svydata}{\itemize{
\item allows access to svrepdesign object's associated data. withReplicates expects it, but we do not use it. We use standata instead
}}

\item{stanmod}{\itemize{
\item the compiled stan model to be passed to rstan::sampling
}}

\item{standata}{\itemize{
\item list of data inputs to be passed to rstan::samlping
}}

\item{par_hat}{\itemize{
\item single set of unconstrained parameters (or estimates like posterior mean) to evaluate the gradient at
}}
}
\value{
the gradient of the log posterior evaluated at par_hat
}
\description{
grad_par
}
