% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_sampling.r
\name{list_2D_row_subset}
\alias{list_2D_row_subset}
\title{list_2D_row_subset}
\usage{
list_2D_row_subset(nmlist, rindex)
}
\arguments{
\item{nmlist}{\itemize{
\item a named list whose elements are 1D or 2D arrays with the same number of rows
}}

\item{rindex}{\itemize{
\item the row index to subset
}}
}
\value{
a named list of types as the input, with elements subset to rows index provided
}
\description{
A helper function to convert a list of MCMC output from Stan to input comparable with the unconstrained transformation.
}
