% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wt_multi_model.R
\name{load_wt_multi_model}
\alias{load_wt_multi_model}
\title{load_wt_multi_model}
\usage{
load_wt_multi_model()
}
\value{
A string containing a Stan model for weighted multinomial regression
}
\description{
load_wt_multi_model
}
