% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_sampling.r
\name{plot.cs_sampling}
\alias{plot.cs_sampling}
\title{plot.cs_sampling}
\usage{
\method{plot}{cs_sampling}(x, varnames = NULL)
}
\arguments{
\item{x}{\itemize{
\item object of type cs_sampling
}}

\item{varnames}{\itemize{
\item vector of names of subset of variable for pairs plotting
}}
}
\description{
plot.cs_sampling
}
