% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_sampling.r
\name{cs_sampling_brms}
\alias{cs_sampling_brms}
\title{cs_sampling_brms}
\usage{
cs_sampling_brms(
  svydes,
  mod,
  data,
  family,
  ctrl_stan = list(chains = 1, iter = 2000, warmup = 1000, thin = 1),
  rep_design = FALSE,
  ctrl_rep = list(replicates = 100, type = "mrbbootstrap")
)
}
\description{
cs_sampling_brms
}
