% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_sampling.r
\name{grad_par}
\alias{grad_par}
\title{grad_par}
\usage{
grad_par(pwts, svydata, stanmod, standata, par_hat)
}
\arguments{
\item{pwts}{\itemize{
\item the weights argument. \code{\link[survey]{withReplicates}} updates this for each set of replicate weights.
}}

\item{svydata}{\itemize{
\item allows access to \code{\link[survey]{svrepdesign}} object's associated data. \code{\link[survey]{withReplicates}} expects it, but we do not use it. We use standata instead
}}

\item{stanmod}{\itemize{
\item the compiled stan model to be passed to \code{\link[rstan]{sampling}}
}}

\item{standata}{\itemize{
\item list of data inputs to be passed to \code{\link[rstan]{sampling}}
}}

\item{par_hat}{\itemize{
\item single set of unconstrained parameters (or estimates like posterior mean) to evaluate the gradient at
}}
}
\value{
the gradient of the log posterior for stanmod evaluated at par_hat
}
\description{
A helper function to use with \code{\link[survey]{withReplicates}}, which estimates the gradient of the log of Stan probability model for a given set of weights.
}
