% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cs_sampling}
\alias{plot.cs_sampling}
\title{plot.cs_sampling}
\usage{
\method{plot}{cs_sampling}(x, varnames = NULL, ...)
}
\arguments{
\item{x}{\itemize{
\item object of type cs_sampling
}}

\item{varnames}{\itemize{
\item optional vector of names of subset of variable for pairs plotting
}}

\item{...}{\itemize{
\item generic parameter
}}
}
\value{
The output of \code{\link[GGally]{ggpairs}}.
}
\description{
Produce pairwise plots comparing each parameter distribution before and after adjustment.
Takes the output list from cs_sampling as input.
}
